+++ src/azure-cli/HISTORY.rst
@@ -73,9 +73,7 @@ Release History
 
 **Cosmos DB**
 
-* Fix missing --type option for deprecation redirections
-* Support az cosmosdb private-endpoint-connection show/delete/approve/reject
-* Support az cosmosdb private-link-resource show
+* Support az cosmosdb private-link-resource and private-endpoint-connection
 
 **Docker**